Add an optional getObjectPropertyWidgetMapping() method.
Since most edit pages want to load and save teh same set of properties based on the widgets, add a method that can be used to create the array, and if that array is set us it when loading and updating the object. Should cut down on repeated code, and subclassed methods almost all of the time.