Allow site 14, remove addToSearchQueue
This upgrades academy to site version 14, which gets rid of all mentions
of nategosearch. It also deprecates the addToSearchQueue method, so all
references to that method have also been removed.

AdminObjectEdit is the first in its inheritance hierarchy to declare
addToSearchQueue, so this will require a major release, as it creates a
breaking change in the interface of AdminObjectEdit.